Changes 9/26
Overwrote custom linker script with auto generated script.  This fixed
errors with the PC jumping to illegal addresses and crashing the
processor.  Removed explicit link to .elf file in PlanAhead run which
allows the .elf file to be programmed to BRAM by the SDK and saves
significant build time.  The GDB debugger in the SDK is incapable of
running, complaining that the board fails to respond.  This error
persists even after fixing linker script.  The algorithm now runs but
needs revision.  Temperature values are readable for each of the
fourteen oscillator modules.